Add cli argument to specify path to run tool
Previously path to root directory to execute tool is specified via
flag -p.

This is cumbersome as it is something that the user expects to be
able to control. Specifying via flag may not allow for tab
autocompletion feature in shell.

Add an option to specify path via cli argument.